feat(transaction): Refactor export transaction functionality to use XLSX library and streamline filter handling
